% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denovo_mutation_rates.R
\name{access_denovo_mutation_rates_sheet}
\alias{access_denovo_mutation_rates_sheet}
\title{Helper Function 3 - get sheet 2 in mutation supplementary file xls}
\usage{
access_denovo_mutation_rates_sheet()
}
\value{
A df from the xls mutations file from sheet 2 only
}
\description{
Helper Function 3 - get sheet 2 in mutation supplementary file xls
}
