% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_sequence_annotations.R
\name{access_feature_page}
\alias{access_feature_page}
\title{Helper function 2 - access biomart feature page}
\usage{
access_feature_page()
}
\value{
A df with ensemble_gene_id, and the \%GC
}
\description{
This is a helper function. Accesses the biomart feature page to get \%GC.
}
