% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_sequence_annotations.R
\name{access_structure_page}
\alias{access_structure_page}
\title{Helper function 1 - access biomart structure page}
\usage{
access_structure_page()
}
\value{
A df with ensemble_gene_id, and the various length annotations
}
\description{
This is a helper function. Accesses the structure feature page to get various length annotations.
}
