% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_protein_data}
\alias{all_protein_data}
\title{all_protein_data}
\format{
A data frame with ... rows and ... variables:
}
\source{
Data accessed with the API from proteomics.db. The data was downloaded on 20MAY25.
}
\usage{
all_protein_data
}
\description{
all_protein_data
}
\keyword{datasets}
