% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_essentiality}
\alias{cell_essentiality}
\title{cell_essentiality}
\format{
A data frame with ... rows and ... variables:
}
\source{
HPSC cell essentiality data from https://www.cell.com/cell-reports/fulltext/S2211-1247(19)30212-8
}
\usage{
cell_essentiality
}
\description{
cell_essentiality
}
\keyword{datasets}
