% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{create_binary_features_subcell}
\alias{create_binary_features_subcell}
\title{Helper Function 4 - create binary features for subcell}
\usage{
create_binary_features_subcell(data, col_names, remove_parentheses_cols = NULL)
}
\value{
A dataframe with subcellular locations from protein atlas in binary format
}
\description{
Helper Function 4 - create binary features for subcell
}
