% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{create_category_features}
\alias{create_category_features}
\title{Helper Function 1.4 - protein expression, creating binary features}
\usage{
create_category_features(data)
}
\value{
A dataframe with protein expression mapped to hgnc_id, brain tissues only
}
\description{
Helper Function 1.4 - protein expression, creating binary features
}
