% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{create_tissue_celltype_level_scores}
\alias{create_tissue_celltype_level_scores}
\title{Helper Function 5 - protein expression, tissue cell type scores}
\usage{
create_tissue_celltype_level_scores(data)
}
\value{
A dataframe with protein expression mapped to hgnc_id, tissue cell type scoers
}
\description{
Helper Function 5 - protein expression, tissue cell type scores
}
