% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddg2p_genes}
\alias{ddg2p_genes}
\title{ddg2p_genes}
\format{
A data frame with ... rows and ... variables:
}
\source{
ddg2p genes Sheet was downloaded from https://ftp.ebi.ac.uk/pub/databases/gene2phenotype/G2P_data_downloads_archive/2025_01_28/DDG2P_2025-01-28.csv.gz on Wednesday 12 March 2025.
}
\usage{
ddg2p_genes
}
\description{
ddg2p_genes
}
\keyword{datasets}
