% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_gene_interactions.R
\name{download_chemical_gene_interactions}
\alias{download_chemical_gene_interactions}
\title{Helper Function 1 - downlaod chemical gene interactions}
\usage{
download_chemical_gene_interactions()
}
\value{
Df of the raw chemical gene interactions data
}
\description{
Helper Function 1 - downlaod chemical gene interactions
}
