% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{download_genetrek}
\alias{download_genetrek}
\title{Helper Function 1 - download genetrek file}
\usage{
download_genetrek()
}
\value{
Df of genetrek table.
}
\description{
This function performs a specific task. Getting the latest genetrek file.
}
