% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denovo_mutation_rates.R
\name{download_mutation_supp_file}
\alias{download_mutation_supp_file}
\title{Helper Function 1 - download mutation supplementary file xls}
\usage{
download_mutation_supp_file()
}
\value{
A temp xls file
}
\description{
Helper Function 1 - download mutation supplementary file xls
}
