% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{download_normal_prot_expr}
\alias{download_normal_prot_expr}
\title{Helper Function 1.1 - download protein expression}
\usage{
download_normal_prot_expr()
}
\value{
A dataframe with protein expression from protein atlas
}
\description{
Helper Function 1.1 - download protein expression
}
