% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{download_subcell_location_file}
\alias{download_subcell_location_file}
\title{Helper Function 1 - download subcellular locations from protein atlas}
\usage{
download_subcell_location_file()
}
\value{
A dataframe with subcellular locations from protein atlas
}
\description{
Helper Function 1 - download subcellular locations from protein atlas
}
