% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exomiser_pheno_score}
\alias{exomiser_pheno_score}
\title{exomiser_pheno_score}
\format{
A data frame with ... rows and ... variables:
}
\source{
Generated with Exomiser and NDD top 5 HPO terms, no human disease data.
}
\usage{
exomiser_pheno_score
}
\description{
exomiser_pheno_score
}
\keyword{datasets}
