% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{filter_and_remove_columns}
\alias{filter_and_remove_columns}
\title{Helper Function 2 - protein expression, filter and remove columns}
\usage{
filter_and_remove_columns(data, cols_to_remove)
}
\value{
A dataframe with protein expression mapped to hgnc_id, filtered columns
}
\description{
Helper Function 2 - protein expression, filter and remove columns
}
