% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{final_pcg_protein_expression}
\alias{final_pcg_protein_expression}
\title{Helper Function 1.7 - protein expression, final df}
\usage{
final_pcg_protein_expression(
  df1_pcg,
  df2_filtered_removed_columns,
  df3_reduced_one_row
)
}
\value{
A dataframe with protein expression mapped to hgnc_id, final df
}
\description{
Helper Function 1.7 - protein expression, final df
}
