% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_constraint_metrics}
\alias{gene_constraint_metrics}
\title{gene_constraint_metrics}
\format{
A data frame with ... rows and ... variables:
}
\source{
Other gene constraint metrics: Shet, DOMINO, and SCoNes
}
\usage{
gene_constraint_metrics
}
\description{
gene_constraint_metrics
}
\keyword{datasets}
