% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphamissense.R
\name{get_alphamissense}
\alias{get_alphamissense}
\title{Main function - get alphamissense gene scores}
\usage{
get_alphamissense(genes)
}
\arguments{
\item{genes}{A df with at least one column with hgnc_id}
}
\value{
A df with hgnc_id and the max alphamissense score for that gene
}
\description{
This is the main function. Gets alphamissense gene sequence scores from
Zenodo and adds HGNC IDs
}
