% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_australia_panelapp}
\alias{get_australia_panelapp}
\title{Main Function 3 - df of Australia Panels NDD}
\usage{
get_australia_panelapp(protein_coding_genes)
}
\value{
Df of the Australia panels NDD
}
\description{
Main Function 3 - df of Australia Panels NDD
}
