% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_australia_panelapp_amber}
\alias{get_australia_panelapp_amber}
\title{Main Function 6 - amber conf aus panelapp}
\usage{
get_australia_panelapp_amber(protein_coding_genes)
}
\value{
Df of the Australia panels NDD. Amber confidence.
}
\description{
Main Function 6 - amber conf aus panelapp
}
