% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_australia_panelapp_amber_cardiac}
\alias{get_australia_panelapp_amber_cardiac}
\title{Main Function 5 - df of Australia cardiac panels amber}
\usage{
get_australia_panelapp_amber_cardiac(protein_coding_genes)
}
\value{
Df of austalia cardiac panels amber.
}
\description{
Main Function 5 - df of Australia cardiac panels amber
}
