% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_australia_panelapp_cardiac}
\alias{get_australia_panelapp_cardiac}
\title{Main Function 2 - df of Australia cardiac panels}
\usage{
get_australia_panelapp_cardiac(protein_coding_genes)
}
\value{
Df of austalia cardiac panels.
}
\description{
Main Function 2 - df of Australia cardiac panels
}
