% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_australia_panelapp_red}
\alias{get_australia_panelapp_red}
\title{Main Function 7 - red conf aus panelapp}
\usage{
get_australia_panelapp_red(protein_coding_genes)
}
\value{
Df of the Australia panels NDD. Red confidence.
}
\description{
Main Function 7 - red conf aus panelapp
}
