% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_australia_panelapp_red_cardiac}
\alias{get_australia_panelapp_red_cardiac}
\title{Main Function 6 - df of Australia cardiac panels red}
\usage{
get_australia_panelapp_red_cardiac(protein_coding_genes)
}
\value{
Df of austalia cardiac panels red.
}
\description{
Main Function 6 - df of Australia cardiac panels red
}
