% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_cell_line_essentiality.R
\name{get_cancer_cell_line_essentiality}
\alias{get_cancer_cell_line_essentiality}
\title{Main Function - get the cancer cell line essentiality of the models of interest.}
\usage{
get_cancer_cell_line_essentiality(
  genes,
  models,
  CRISPRGeneEffects,
  get_mean = FALSE
)
}
\arguments{
\item{genes}{A df with hgnc_ids column}

\item{models}{A df of the model ids of interest, obtained from get_cancer_cell_line_models()}

\item{CRISPRGeneEffects}{The url to the CRISPRGeneEffects file, obtained from get_latest_depmap_release()}

\item{get_mean}{The option to get the means of the cell lines instead of the raw essentiality scores.
Default is FALSE.}
}
\value{
Df of the essentiality scores. Either the raw scores or the mean, depending on get_mean.
}
\description{
This is the main function. Getting cell line essentiality scores of the models of interest.
}
