% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_cell_line_essentiality.R
\name{get_cancer_cell_line_models}
\alias{get_cancer_cell_line_models}
\title{Helper Function 3 - get model ids of the models of interest.}
\usage{
get_cancer_cell_line_models(vector_of_OncotreeLineage, model_url)
}
\arguments{
\item{vector_of_OncotreeLineage}{A vector of one or more model lineages. These are the available lineages:
Ovary/Fallopian Tube, Myeloid, Bowel, Skin, Bladder/Urinary Tract, Lung, Kidney
Breast, Lymphoid, Pancreas, CNS/Brain, Soft Tissue, Bone, Fibroblast
Esophagus/Stomach, Thyroid, Peripheral Nervous System, Pleura, Prostate
Biliary Tract, Head and Neck, Uterus, Ampulla of Vater, Liver, Cervix, Eye
Vulva/Vagina, Adrenal Gland, Testis, Other, Normal, NA, Muscle, Embryonal, Hair}
}
\value{
Df of the model ids.
}
\description{
This function performs a specific task. Getting the model ids of the lineage decided by the input.
}
