% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chemical_gene_interactions.R
\name{get_chemical_gene_interactions}
\alias{get_chemical_gene_interactions}
\title{Main Function 1 - gets chemical gene interactions for each gene}
\usage{
get_chemical_gene_interactions(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{Needs to be a df with hgnc_id and entrez_id columns}
}
\value{
Df of the chemical gene interactions data joined by entrez_id
}
\description{
Main Function 1 - gets chemical gene interactions for each gene
}
