% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_ddg2p}
\alias{get_ddg2p}
\title{Main Function 11 - df of ddg2p}
\usage{
get_ddg2p(protein_coding_genes)
}
\value{
df of ddg2p. Definitive and strong confidence. Brain/Cognition only.
}
\description{
Main Function 11 - df of ddg2p
}
