% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_ddg2p_limited}
\alias{get_ddg2p_limited}
\title{Main Function 11 - df of ddg2p limited}
\usage{
get_ddg2p_limited(protein_coding_genes)
}
\value{
df of ddg2p. Moderate confidence. Brain/Cognition only.
}
\description{
Main Function 11 - df of ddg2p limited
}
