% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_ddg2p_moderate}
\alias{get_ddg2p_moderate}
\title{Main Function 12 - df of ddg2p moderate}
\usage{
get_ddg2p_moderate(protein_coding_genes)
}
\value{
df of ddg2p. Moderate confidence. Brain/Cognition only.
}
\description{
Main Function 12 - df of ddg2p moderate
}
