% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denovo_mutation_rates.R
\name{get_denovo_mutation_rates}
\alias{get_denovo_mutation_rates}
\title{Main Function - get denovo mutation rates}
\usage{
get_denovo_mutation_rates(protein_coding_genes)
}
\value{
A dataframe with protein expression mapped to hgnc_id
}
\description{
Main Function - get denovo mutation rates
}
