% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_gel_panelapp}
\alias{get_gel_panelapp}
\title{Main Function 2 - df of GEL panels NDD}
\usage{
get_gel_panelapp(protein_coding_genes)
}
\value{
A df of the GEL panels for NDD.
}
\description{
Main Function 2 - df of GEL panels NDD
}
