% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_gel_panelapp_amber}
\alias{get_gel_panelapp_amber}
\title{Main Function 4 - amber conf gel panelapp}
\usage{
get_gel_panelapp_amber(protein_coding_genes)
}
\value{
A df of the GEL panels for NDD. Amber confidence.
}
\description{
Main Function 4 - amber conf gel panelapp
}
