% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_gel_panelapp_amber_cardiac}
\alias{get_gel_panelapp_amber_cardiac}
\title{Main Function 2 - df GEL panels cardiac amber}
\usage{
get_gel_panelapp_amber_cardiac(protein_coding_genes)
}
\value{
Df of GEL cardiac panels amber.
}
\description{
Main Function 2 - df GEL panels cardiac amber
}
