% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_gel_panelapp_cardiac}
\alias{get_gel_panelapp_cardiac}
\title{Main Function 1 - df GEL panels cardiac}
\usage{
get_gel_panelapp_cardiac(protein_coding_genes)
}
\value{
Df of GEL cardiac panels.
}
\description{
Main Function 1 - df GEL panels cardiac
}
