% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_gel_panelapp_red}
\alias{get_gel_panelapp_red}
\title{Main Function 5 - red conf gel panelapp}
\usage{
get_gel_panelapp_red(protein_coding_genes)
}
\value{
A df of the GEL panels for NDD. Red confidence.
}
\description{
Main Function 5 - red conf gel panelapp
}
