% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R
\name{get_gel_panelapp_red_cardiac}
\alias{get_gel_panelapp_red_cardiac}
\title{Main Function 3 - df GEL panels cardiac red}
\usage{
get_gel_panelapp_red_cardiac(protein_coding_genes)
}
\value{
Df of GEL cardiac panels red.
}
\description{
Main Function 3 - df GEL panels cardiac red
}
