% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expression.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Main function - get gene expression of the organ and organism of interest}
\usage{
get_gene_expression(protein_coding_genes, organism, vector_of_organs)
}
\arguments{
\item{protein_coding_genes}{A df with at least one column of hgnc_id and ensembl_gene_id}

\item{organism}{A string of either 'mouse' or 'human'.}

\item{organ}{A vector of strings of the organs of interest.}
}
\value{
A data frame with the columns of the organ of interest of either mouse or human of gene expression.
}
\description{
This is the main functions. Gets the gene expression RPKM of the organ and
organism of interest. Either mouse or human.
}
