% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_sequence_annotations.R
\name{get_gene_sequence_annotations}
\alias{get_gene_sequence_annotations}
\title{Main function - get gene sequence annotations from biomart}
\usage{
get_gene_sequence_annotations(protein_coding_genes, keep_version = FALSE)
}
\arguments{
\item{protein_coding_genes}{A df with at least hgnc_id and ensembl_gene_id}
}
\value{
A df with hgnc_id, ensemble_gene_id, and the gene sequence annotations
Specifically: gene length, transcript length, cds length, \%GC, and version of the ensembl entry
}
\description{
This is the main function. Gets gene sequence annotations from biomart
}
