% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_genetrek}
\alias{get_genetrek}
\title{Main Function 2 - cleans genetrek file and gives you results for the genes specified}
\usage{
get_genetrek(protein_coding_genes, columns = NULL)
}
\value{
Df of genetrek for specified genes.
}
\description{
Main Function 2 - cleans genetrek file and gives you results for the genes specified
}
