% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gnomad_constraint.R
\name{get_gnomad_constraint}
\alias{get_gnomad_constraint}
\title{Main function - get Gnomad constraint scores}
\usage{
get_gnomad_constraint(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{A df with at least hgnc_id and ensembl_gene_id}
}
\value{
A data frame with hgnc_id, ensemble_gene_id, and the gnomad constraint scores
}
\description{
This is the main function. Gets Gnomad contraints scores from v4.1
}
