% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hpsc_essentiality.R
\name{get_hpsc_essentiality}
\alias{get_hpsc_essentiality}
\title{Main function - get hPSC essentiality.}
\usage{
get_hpsc_essentiality(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{A data frame with HGNC ids.}
}
\value{
Data frame of hPSC essentiality scores.
}
\description{
This is the main function. Gets the hPSC essentiality scores from RData file.
https://www.cell.com/cell-reports/fulltext/S2211-1247(19)30212-8
}
