% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_cell_line_essentiality.R
\name{get_latest_depmap_release}
\alias{get_latest_depmap_release}
\title{Helper Function 1 - get latest DepMap release links}
\usage{
get_latest_depmap_release()
}
\value{
List of links to the latest DepMap release of Model.csv and CRISPRGeneEffect.csv
}
\description{
This function performs a specific task. Getting the latest DepMap release links.
}
