% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cancer_cell_line_essentiality.R
\name{get_lineage}
\alias{get_lineage}
\title{Helper Function 2 - get latest DepMap Model.csv as a df.}
\usage{
get_lineage(file_url, string)
}
\arguments{
\item{file_url}{Url of Model.csv, the ouput of get_latest_depmap_release() $model}

\item{string}{A string of the model of interest.}
}
\value{
Df of the latest DepMap release of Model.csv.
}
\description{
This function performs a specific task. Getting the latest DepMap Model.csv as a df.
}
