% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mouse_viability.R
\name{get_mouse_protein_coding_genes}
\alias{get_mouse_protein_coding_genes}
\title{Helper Function 1 - get mouse protein coding genes}
\usage{
get_mouse_protein_coding_genes()
}
\value{
Large character. Of MGI ids.
}
\description{
This function performs a specific task. Getting mouse protein coding genes.
}
