% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mouse_viability.R
\name{get_mouse_viability}
\alias{get_mouse_viability}
\title{Main Function}
\usage{
get_mouse_viability(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{Output of get_protein_coding_genes()}
}
\value{
A data frame with the combined mouse viability.
}
\description{
This is the main function which uses helper functions to obtain the combined mouse viability.
}
