% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_expression.R
\name{get_organ}
\alias{get_organ}
\title{Helper Function 1 - get organ columns}
\usage{
get_organ(organism, organ)
}
\arguments{
\item{organism}{A string of either 'mouse' or 'human'.}

\item{organ}{A string of the organ of interest.}
}
\value{
A data frame with the columns of the organ of interest of either mouse or human.
}
\description{
This function performs a specific task. Gets the organ of interest columns
from the mouse or human gene expression data.
https://apps.kaessmannlab.org/evodevoapp/
mouse IDs are different. add script to match to human orthologs.
}
