% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_other_constraint.R
\name{get_other_constraint}
\alias{get_other_constraint}
\title{Main function - get Shet, DOMINO, and SCoNes constraint}
\usage{
get_other_constraint(protein_coding_genes)
}
\arguments{
\item{protein_coding_genes}{A df with at least one column with hgnc_id.}
}
\value{
A df with hgnc_id and the Shet, DOMINO, and SCoNes scores.
}
\description{
This is the main function. Reads the RData file with the Shet, DOMINO, and
SCoNes scores as these are static.
}
