% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_coding_genes.R
\name{get_protein_coding_genes}
\alias{get_protein_coding_genes}
\title{Get all protein coding genes data}
\usage{
get_protein_coding_genes()
}
\value{
A dataframe with protein coding gene data from EBI database.
}
\description{
This function gets the df from EBI of all protein coding genes.
}
