% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{get_protein_expression_hpa}
\alias{get_protein_expression_hpa}
\title{Main Function 1 - get protein expression from HPA}
\usage{
get_protein_expression_hpa(protein_coding_genes, brain_tissues_only = TRUE)
}
\value{
A dataframe with protein expression from HPA mapped to hgnc_id
}
\description{
Main Function 1 - get protein expression from HPA
}
