% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{get_protein_expression_pdb}
\alias{get_protein_expression_pdb}
\title{Main Function 2 - get protein expression from ProteomicsDB}
\usage{
get_protein_expression_pdb(
  protein_coding_genes,
  specific_tissue = NULL,
  zeros = TRUE
)
}
\arguments{
\item{protein_coding_genes}{A dataframe with protein coding genes}

\item{specific_tissue}{A string with the name of the tissue to filter for}

\item{zeros}{A boolean indicating whether to fill missing values with 0}
}
\value{
A dataframe with protein expression from HPA mapped to hgnc_id
}
\description{
Main Function 2 - get protein expression from ProteomicsDB
}
