% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{get_subcellular_location}
\alias{get_subcellular_location}
\title{Main Function - get subcellular locations}
\usage{
get_subcellular_location(protein_coding_genes, high_support_and_binary = TRUE)
}
\value{
A dataframe with subcellular locations mapped to hgnc_id
}
\description{
keeping all reliability levels.
}
