% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_sysndd}
\alias{get_sysndd}
\title{Main Function 8 - df of SysNDD definitive}
\usage{
get_sysndd(protein_coding_genes)
}
\value{
df of SysNDD. Definitive confidence.
}
\description{
Main Function 8 - df of SysNDD definitive
}
