% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_sysndd_limited}
\alias{get_sysndd_limited}
\title{Main Function 10 - df of SysNDD limited}
\usage{
get_sysndd_limited(protein_coding_genes)
}
\value{
df of SysNDD. Limited confidence.
}
\description{
Main Function 10 - df of SysNDD limited
}
