% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{get_sysndd_moderate}
\alias{get_sysndd_moderate}
\title{Main Function 9 - df of SysNDD moderate}
\usage{
get_sysndd_moderate(protein_coding_genes)
}
\value{
df of SysNDD. Moderate confidence.
}
\description{
Main Function 9 - df of SysNDD moderate
}
