% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_symbol_checker.R
\name{hgnc_checker}
\alias{hgnc_checker}
\title{Check HGNC Gene Symbols}
\usage{
hgnc_checker(
  gene_symbols,
  file_path = "gene_with_protein_product.txt",
  download_url =
    "https://storage.googleapis.com/public-download-files/hgnc/tsv/tsv/locus_types/gene_with_protein_product.txt"
)
}
\arguments{
\item{gene_symbols}{A vector of gene symbols to check.}

\item{file_path}{A file path to the HGNC data file. If the file does not exist, it will be downloaded.}

\item{download_url}{URL to download the HGNC data file if it is not present.}
}
\value{
A data frame with columns \code{hgnc_id}, \code{gene_symbol}, and \code{type}.
}
\description{
This function checks HGNC gene symbols against a database of protein-coding genes
and returns matching information.
}
