% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{high_support_locations}
\alias{high_support_locations}
\title{Helper Function 2 - keep subcell columns of high support}
\usage{
high_support_locations(subcellular_location)
}
\value{
A dataframe with subcellular locations from protein atlas with high support
}
\description{
Helper Function 2 - keep subcell columns of high support
}
