% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiphive}
\alias{hiphive}
\title{NDD_HiPhive_Prioritiser_genes}
\format{
A data frame with ... rows and ... variables:
}
\source{
Generated with Exomiser and NDD top 5 HPO terms.
}
\usage{
hiphive
}
\description{
NDD_HiPhive_Prioritiser_genes
}
\keyword{datasets}
