% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human}
\alias{human}
\title{human}
\format{
A data frame with ... rows and ... variables:
}
\source{
Human gene expression data from https://apps.kaessmannlab.org/evodevoapp/
}
\usage{
human
}
\description{
human
}
\keyword{datasets}
