% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lethal_terms}
\alias{lethal_terms}
\title{lethal_terms}
\format{
A data frame with ... rows and ... variables:
}
\source{
lethal_terms data from Dickinson et al.(PMID: 27626380)
}
\usage{
lethal_terms
}
\description{
lethal_terms
}
\keyword{datasets}
