% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ndd_input_genes.R
\name{load_and_clean_panel}
\alias{load_and_clean_panel}
\title{Helper Function 1 - load and clean panels}
\usage{
load_and_clean_panel(url)
}
\value{
cleaning of the panel.
}
\description{
This function performs a specific task. Getting the latest genetrek file.
}
