% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_features.R
\name{merge_features}
\alias{merge_features}
\title{Main Function - merging features}
\usage{
merge_features(df_list)
}
\arguments{
\item{df_list}{List of dataframes to merge.}
}
\value{
A dataframe with hgnc_id and all the features.
}
\description{
Merging data/features from multiple sources.
}
