% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{merge_prefixed_features}
\alias{merge_prefixed_features}
\title{Helper Function 5 - merge prefixed features in subcell}
\usage{
merge_prefixed_features(data)
}
\value{
A dataframe with subcellular locations from protein atlas merged prefixes
}
\description{
Helper Function 5 - merge prefixed features in subcell
}
