% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{mouse}
\format{
A data frame with ... rows and ... variables:
}
\source{
Mouse gene expression data from https://apps.kaessmannlab.org/evodevoapp/
}
\usage{
mouse
}
\description{
mouse
}
\keyword{datasets}
