% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{protein_expression_pdb_tissues}
\alias{protein_expression_pdb_tissues}
\title{Helper Function 2.2 - show all tissues in protein expression raw file}
\usage{
protein_expression_pdb_tissues()
}
\value{
A clean dataframe with all tissues in protein expression from ProteomicsDB
}
\description{
Helper Function 2.2 - show all tissues in protein expression raw file
}
