% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphamissense.R
\name{read_alphamissense_gene_hg38}
\alias{read_alphamissense_gene_hg38}
\title{Helper Function 1 - read alphamissense file from zenodo}
\usage{
read_alphamissense_gene_hg38()
}
\value{
Dataframe of alphamissense gene hg38 scores.
}
\description{
This function performs a specific task. Reads the alphamissense gene hg38
scores file from zenodo.
}
