% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_denovo_mutation_rates.R
\name{read_sheet}
\alias{read_sheet}
\title{Helper Function 2 - read mutation supplementary file xls}
\usage{
read_sheet(temp_file, sheet_number)
}
\value{
A df from the xls mutations file
}
\description{
Helper Function 2 - read mutation supplementary file xls
}
