% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{reduce_to_one_row}
\alias{reduce_to_one_row}
\title{Helper Function 1.6 - protein expression, distinct rows}
\usage{
reduce_to_one_row(data, id_col = "hgnc_id")
}
\value{
A dataframe with protein expression mapped to hgnc_id, tissue cell type scores, distinct rows
}
\description{
Helper Function 1.6 - protein expression, distinct rows
}
