% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subcellular_location.R
\name{remove_parentheses}
\alias{remove_parentheses}
\title{Helper Function 3 - subcellular locations remove parenthesis}
\usage{
remove_parentheses(data, col_name)
}
\value{
A dataframe with subcellular locations from protein atlas with parenthesis removed
}
\description{
Helper Function 3 - subcellular locations remove parenthesis
}
