% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{select_brain_tissues}
\alias{select_brain_tissues}
\title{Helper Function 3 - protein expression, select brain tissues}
\usage{
select_brain_tissues(data)
}
\value{
A dataframe with protein expression mapped to hgnc_id, brain tissues only
}
\description{
Helper Function 3 - protein expression, select brain tissues
}
