% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sysndd_genes}
\alias{sysndd_genes}
\title{sysndd_genes}
\format{
A data frame with ... rows and ... variables:
}
\source{
sysndd genes Sheet was downloaded from https://sysndd.dbmr.unibe.ch/Genes on Wednesday 12 March 2025. Website version v0.1.0.
}
\usage{
sysndd_genes
}
\description{
sysndd_genes
}
\keyword{datasets}
