% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardiac_genes.R, R/get_ndd_input_genes.R
\name{get_australia_panelapp_red}
\alias{get_australia_panelapp_red}
\title{Main Function 6 - df of Australia cardiac panels red}
\usage{
get_australia_panelapp_red(protein_coding_genes)

get_australia_panelapp_red(protein_coding_genes)
}
\value{
Df of austalia cardiac panels red.

Df of the Australia panels NDD. Red confidence.
}
\description{
Main Function 6 - df of Australia cardiac panels red

Main Function 7 - red conf aus panelapp
}
