% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mouse_viability.R
\name{get_impc_viability}
\alias{get_impc_viability}
\title{Helper Function 2 - get IMPC viability.}
\usage{
get_impc_viability()
}
\value{
Dataframe of IMPC viability.
\describe{
\item{mgi_id}{MDI IF for gene}
\item{viability_impc}{IMPC viability}
}
}
\description{
This function performs a specific task. Getting IMPC viability.
}
