% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_expression.R
\name{get_protein_expression}
\alias{get_protein_expression}
\title{Main Function - get protein expression}
\usage{
get_protein_expression(protein_coding_genes, brain_tissues_only = TRUE)
}
\value{
A dataframe with protein expression mapped to hgnc_id
}
\description{
Main Function - get protein expression
}
